backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pinecone-client==3.0.0
sentence-transformers==2.2.2
requests==2.31.0